dj_client/requirements.txt
Django==3.2.13
djangogrpcframework==0.2.1
djangorestframework==3.13.1
grpcio==1.44.0
grpcio-tools==1.44.0